Added fp code samples
This is a basic implementation of a lisp-style list and related
functional-programming algorithms: map, reduce, filter, length, slice,
insert, deleteRange.

None of this code has been tested (obviously), but it should probably
mostly work (unit tests and hell, even just attempting to compile tbd);
this does also also include a simple ref-counted object system (c++),
so yeah... lots of room for bugs.

The eventual goal here will be to a) build + test this (probably move
to a separate project for that), and b) translate + reimplement in
assembly.

Having a lisp-like object system in asm, even if I don't have a full
frontend parser + repl, would be neat.

And this would probably also include a bignum, symbolic math, and
matrix library, since those are other projects I'm also working on...