rewrote memcpy to use movs instructions
Note: this entire function could probably be rewritten as

memcpy:
    cld
    movsb
    ret

What my code does is try to use movsd (instead of movsb), at the cost
of 3 additional branches. This is probably not worth it b/c a)
pipelining, b) the real cost here is memory access, not clock cycles,
and c) movsb should be heavily optimized anyways.

The difference is probably negligible, but it would be interesting to
do a performance comparison between this, movsb, and the original
algorithm which just moves stuff between registers (prediction: movsb
will be faster in all cases).